Making work channels replenish themselves
Previously we were having issues with work channels dying (if, for
example, a request publisher was no longer alive) and therefore not
being able to continue with any other work.

In the case of a failure, we now lock the work channel on death, recreate
it and then unlock again for easier access.

Realistically, this shouldn’t be something that happens too often, so
the slight performance kick for the recreations is a small price at the
moment.